Lab_2/requirements.txt
astroid==2.3.3
Django==2.2.6
django-chartjs==1.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
uWSGI==2.0.18
wrapt==1.11.2
